requirements.txt
numpy
pandas
scipy
networkx
pyyaml
tqdm
matplotlib
torch
torchvision
torchaudio
torch-geometric
requests
tensorboard